translation-service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
aiohttp==3.10.5
python-multipart==0.0.9
PyMuPDF==1.24.9
python-docx==1.1.2
boto3==1.35.18